requirements.txt
annotated-types==0.5.0
anyio==3.7.1
click==8.1.8
colorama==0.4.6
exceptiongroup==1.2.2
fastapi==0.103.2
greenlet==3.1.1
h11==0.14.0
idna==3.10
importlib-metadata==6.7.0
iniconfig==2.0.0
itsdangerous==2.1.2
MarkupSafe==2.1.5
mysql-connector-python==8.0.33
packaging==24.0
pluggy==1.2.0
protobuf==3.20.3
pydantic==2.5.3
pydantic_core==2.14.6
pytest==7.4.4
python-dotenv==0.21.1
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.27.0
tomli==2.0.1
typing_extensions==4.7.1
uvicorn==0.22.0
zipp==3.15.0